+++ R/profile.R
@@ -98,13 +98,13 @@ gcs_first <- function(bucket = Sys.getenv("GCS_SESSION_BUCKET")){
                              return()
                            })
       }
+      
+      set_scopes()
 
-      options(googleAuthR.scopes.selected =
-                "https://www.googleapis.com/auth/devstorage.read_write")
       auth_try <- googleAuthR::gar_gce_auth()
       if(is.null(auth_try)){
         message("GCE auth didn't work, looking for GCS_AUTH_FILE")
-        gcs_auth()
+        gcs_auth(Sys.getenv("GCS_AUTH_FILE"))
       }
 
       o <- tryCatch(gcs_list_objects(prefix = gcs_rdata, bucket = bucket),
@@ -181,11 +181,11 @@ gcs_last <- function(bucket = Sys.getenv("GCS_SESSION_BUCKET")){
             no attempt to save workspace")
     return()
   }
-  options(googleAuthR.scopes.selected = "https://www.googleapis.com/auth/devstorage.read_write")
+  set_scopes()
   auth_try <- gar_gce_auth()
   if(is.null(auth_try)){
-    message("GCE auth didn't work, looking for GCS_AUTH")
-    gcs_auth()
+    message("GCE auth didn't work, looking for GCS_AUTH_FILE")
+    gcs_auth(Sys.getenv("GCS_AUTH_FILE"))
   }
 
   message("\nSaving data to Google Cloud Storage:\n",